requirements.txt
pytesseract==0.2.6
latex2sympy3==0.1.1
kivy.deps.sdl2==0.1.18
sympy==1.3
Pillow==5.4.1
kivy==1.10.1
